Rename eups module to eups_module
