+++ python/lsst/ci/eups_module.py
